add PipeDuplicator to pipe_utils.hpp
