+++ DirectProgramming/DPC++FPGA/ReferenceDesigns/mvdr_beamforming/src/MVDR.hpp
@@ -6,7 +6,6 @@
 #include <array>
 
 // utility classes
-#include "PipeDuplicator.hpp"
 #include "mvdr_complex.hpp"
 #include "pipe_utils.hpp"
