+++ DirectProgramming/DPC++FPGA/include/pipe_utils.hpp
@@ -15,15 +15,26 @@
 This header defines the following utilities for use with pipes in oneAPI programs.
 
 1. PipeArray
-    template <class Id,          // identifier for the pipe array
-              typename BaseTy,   // type to write/read for each pipe
-              size_t min_depth,  // minimum capacity of each pipe
-              size_t... dims     // depth of each dimension in the array
-                                 // any number of dimensions are supported
-              >
-    struct PipeArray
+     template <class Id,          // identifier for the pipe array
+               typename BaseTy,   // type to write/read for each pipe
+               size_t min_depth,  // minimum capacity of each pipe
+               size_t... dims     // depth of each dimension in the array
+                                  // any number of dimensions are supported
+               >
+     struct PipeArray
 
-2. 
+2. PipeDuplicator 
+     Connect a kernel that writes to a single pipe to multiple pipe instances,
+     each of which will receive the same data.
+     A blocking write will perform a blocking write to each pipe.  A non-blocking
+     write will perform a non-blocking write to each pipe, and set success to
+     true only if ALL writes were successful.
+
+     template <class Id,          // name of this PipeDuplicator
+               typename T,        // data type to transfer
+               typename... Pipes  // all pipes to send duplicated writes to
+               >
+     struct PipeDuplicator
 
 */
 
@@ -83,7 +94,7 @@ struct write_currying<WriteFunc, BaseTy, std::index_sequence<I...>> {
 
 
 // =============================================================
-// Pipe Array
+// PipeArray
 // =============================================================
 
 template <class Id,          // identifier for the pipe array
@@ -164,6 +175,65 @@ struct PipeArray {
   }
 
 };  // end of struct PipeArray
+
+// =============================================================
+// PipeDuplicator
+// =============================================================
+
+// Connect a kernel that writes to a single pipe to multiple pipe instances,
+// each of which will receive the same data.
+// A blocking write will perform a blocking write to each pipe.  A non-blocking
+// write will perform a non-blocking write to each pipe, and set success to
+// true only if ALL writes were successful.
+
+// primary template, dummy
+template <class Id,          // name of this PipeDuplicator
+          typename T,        // data type to transfer
+          typename... Pipes  // all pipes to send duplicated writes to
+          >
+struct PipeDuplicator {};
+
+// recursive case, write to each pipe
+template <class Id,                   // name of this PipeDuplicator
+          typename T,                 // data type to transfer
+          typename FirstPipe,         // at least one output pipe
+          typename... RemainingPipes  // additional copies of the output pipe
+          >
+struct PipeDuplicator<Id, T, FirstPipe, RemainingPipes...> {
+  PipeDuplicator() = delete;  // ensure we cannot create an instance
+
+  // Non-blocking write
+  static void write(const T &data, bool &success) {
+    bool local_success;
+    FirstPipe::write(data, local_success);
+    success = local_success;
+    PipeDuplicator<Id, T, RemainingPipes...>::write(data, local_success);
+    success &= local_success;
+  }
+
+  // Blocking write
+  static void write(const T &data) {
+    FirstPipe::write(data);
+    PipeDuplicator<Id, T, RemainingPipes...>::write(data);
+  }
+};
+
+// base case for recursion, no pipes to write to
+// also useful as a 'null' pipe, writes don't do anything
+template <class Id,   // name of this PipeDuplicator
+          typename T  // data type to transfer
+          >
+struct PipeDuplicator<Id, T> {
+  PipeDuplicator() = delete;  // ensure we cannot create an instance
+
+  // Non-blocking write
+  static void write(const T & /*data*/, bool &success) { success = true; }
+
+  // Blocking write
+  static void write(const T & /*data*/) {
+    // do nothing
+  }
+};
 
 #endif /* __PIPE_ARRAY_HPP__ */
